+++ packages/std/src/exports.rs
@@ -13,7 +13,7 @@ use core::{marker::PhantomData, ptr};
 use serde::de::DeserializeOwned;
 
 use crate::deps::OwnedDeps;
-#[cfg(any(feature = "stargate", feature = "ibcv2"))]
+#[cfg(any(feature = "stargate", feature = "ibc2"))]
 use crate::ibc::IbcReceiveResponse;
 use crate::ibc::{IbcBasicResponse, IbcDestinationCallbackMsg, IbcSourceCallbackMsg};
 #[cfg(feature = "stargate")]
@@ -22,8 +22,8 @@ use crate::ibc::{
     IbcPacketTimeoutMsg,
 };
 use crate::ibc::{IbcChannelOpenMsg, IbcChannelOpenResponse};
-#[cfg(feature = "ibcv2")]
-use crate::ibcv2::IBCv2PacketReceiveMsg;
+#[cfg(feature = "ibc2")]
+use crate::ibc2::Ibc2PacketReceiveMsg;
 use crate::imports::{ExternalApi, ExternalQuerier, ExternalStorage};
 use crate::memory::{Owned, Region};
 use crate::panic::install_panic_handler;
@@ -48,9 +48,9 @@ extern "C" fn requires_staking() {}
 #[no_mangle]
 extern "C" fn requires_stargate() {}
 
-#[cfg(feature = "ibcv2")]
+#[cfg(feature = "ibc2")]
 #[no_mangle]
-extern "C" fn requires_ibcv2() {}
+extern "C" fn requires_ibc2() {}
 
 #[cfg(feature = "cosmwasm_1_1")]
 #[no_mangle]
@@ -526,21 +526,17 @@ where
     Region::from_vec(v).to_heap_ptr() as u32
 }
 
-/// do_ibcv2_packet_receive is designed for use with #[entry_point] to make a "C" extern
+/// do_ibc2_packet_receive is designed for use with #[entry_point] to make a "C" extern
 ///
-/// contract_fn is called when this chain receives an IBCv2 payload on port belonging
+/// contract_fn is called when this chain receives an Ibc2 payload on port belonging
 /// to this contract
 ///
 /// - `Q`: custom query type (see QueryRequest)
 /// - `C`: custom response message type (see CosmosMsg)
 /// - `E`: error type for responses
-#[cfg(feature = "ibcv2")]
-pub fn do_ibcv2_packet_receive<Q, C, E>(
-    contract_fn: &dyn Fn(
-        DepsMut<Q>,
-        Env,
-        IBCv2PacketReceiveMsg,
-    ) -> Result<IbcReceiveResponse<C>, E>,
+#[cfg(feature = "ibc2")]
+pub fn do_ibc2_packet_receive<Q, C, E>(
+    contract_fn: &dyn Fn(DepsMut<Q>, Env, Ibc2PacketReceiveMsg) -> Result<IbcReceiveResponse<C>, E>,
     env_ptr: u32,
     msg_ptr: u32,
 ) -> u32
@@ -550,7 +546,7 @@ where
     E: ToString,
 {
     install_panic_handler();
-    let res = _do_ibcv2_packet_receive(
+    let res = _do_ibc2_packet_receive(
         contract_fn,
         env_ptr as *mut Region<Owned>,
         msg_ptr as *mut Region<Owned>,
@@ -927,13 +923,9 @@ where
     }
 }
 
-#[cfg(feature = "ibcv2")]
-fn _do_ibcv2_packet_receive<Q, C, E>(
-    contract_fn: &dyn Fn(
-        DepsMut<Q>,
-        Env,
-        IBCv2PacketReceiveMsg,
-    ) -> Result<IbcReceiveResponse<C>, E>,
+#[cfg(feature = "ibc2")]
+fn _do_ibc2_packet_receive<Q, C, E>(
+    contract_fn: &dyn Fn(DepsMut<Q>, Env, Ibc2PacketReceiveMsg) -> Result<IbcReceiveResponse<C>, E>,
     env_ptr: *mut Region<Owned>,
     msg_ptr: *mut Region<Owned>,
 ) -> ContractResult<IbcReceiveResponse<C>>
@@ -948,7 +940,7 @@ where
         unsafe { Region::from_heap_ptr(ptr::NonNull::new(msg_ptr).unwrap()).into_vec() };
 
     let env: Env = try_into_contract_result!(from_json(env));
-    let msg: IBCv2PacketReceiveMsg = try_into_contract_result!(from_json(msg));
+    let msg: Ibc2PacketReceiveMsg = try_into_contract_result!(from_json(msg));
 
     let mut deps = make_dependencies();
     contract_fn(deps.as_mut(), env, msg).into()