+++ packages/std/src/ibc2.rs
@@ -8,44 +8,44 @@ use crate::{Addr, Binary, Timestamp};
 #[non_exhaustive]
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
-pub struct IBCv2Payload {
+pub struct Ibc2Payload {
     /// The port id on the chain where the packet is sent from.
     pub source_port: String,
     /// The port id on the chain where the packet is sent to.
     pub destination_port: String,
     /// Version of the receiving contract.
     pub version: String,
-    /// Encoding used to serialize the [IBCv2Payload::value].
+    /// Encoding used to serialize the [Ibc2Payload::value].
     pub encoding: String,
     /// Encoded payload data.
     pub value: Binary,
 }
 
-/// These are messages in the IBC lifecycle using the new IBCv2 approach.
-/// Only usable by IBCv2-enabled contracts
+/// These are messages in the IBC lifecycle using the new Ibc2 approach.
+/// Only usable by Ibc2-enabled contracts
 #[non_exhaustive]
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
-pub enum IBCv2Msg {
-    /// Sends an IBCv2 packet with given payloads over the existing channel.
+pub enum Ibc2Msg {
+    /// Sends an Ibc2 packet with given payloads over the existing channel.
     SendPacket {
         channel_id: String,
         timeout: Timestamp,
-        payloads: Vec<IBCv2Payload>,
+        payloads: Vec<Ibc2Payload>,
     },
 }
 
-/// The message that is passed into `ibcv2_packet_receive`
+/// The message that is passed into `ibc2_packet_receive`
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
 #[non_exhaustive]
-pub struct IBCv2PacketReceiveMsg {
-    pub payload: IBCv2Payload,
+pub struct Ibc2PacketReceiveMsg {
+    pub payload: Ibc2Payload,
     pub relayer: Addr,
     pub source_client: String,
 }
 
-impl IBCv2PacketReceiveMsg {
-    pub fn new(payload: IBCv2Payload, relayer: Addr, source_client: String) -> Self {
+impl Ibc2PacketReceiveMsg {
+    pub fn new(payload: Ibc2Payload, relayer: Addr, source_client: String) -> Self {
         Self {
             payload,
             relayer,
@@ -58,11 +58,11 @@ impl IBCv2PacketReceiveMsg {
 mod tests {
     use serde_json::to_string;
 
-    use crate::IBCv2Payload;
+    use crate::Ibc2Payload;
 
     #[test]
-    fn ibcv2_payload_serialize() {
-        let packet = IBCv2Payload {
+    fn ibc2_payload_serialize() {
+        let packet = Ibc2Payload {
             source_port: "sending-contractr-port".to_string(),
             destination_port: "receiving-contract-port".to_string(),
             version: "v1".to_string(),