+++ packages/vm/src/lib.rs
@@ -32,6 +32,8 @@ pub use crate::calls::{
     call_migrate_with_info_raw, call_query, call_query_raw, call_reply, call_reply_raw, call_sudo,
     call_sudo_raw,
 };
+#[cfg(feature = "ibc2")]
+pub use crate::calls::{call_ibc2_packet_receive, call_ibc2_packet_receive_raw};
 #[cfg(feature = "stargate")]
 pub use crate::calls::{
     call_ibc_channel_close, call_ibc_channel_close_raw, call_ibc_channel_connect,
@@ -39,8 +41,6 @@ pub use crate::calls::{
     call_ibc_packet_ack, call_ibc_packet_ack_raw, call_ibc_packet_receive,
     call_ibc_packet_receive_raw, call_ibc_packet_timeout, call_ibc_packet_timeout_raw,
 };
-#[cfg(feature = "ibcv2")]
-pub use crate::calls::{call_ibcv2_packet_receive, call_ibcv2_packet_receive_raw};
 pub use crate::capabilities::capabilities_from_csv;
 pub use crate::config::{CacheOptions, Config, WasmLimits};
 pub use crate::errors::{